Rename to toggleSubSubscription
